+++ repos/ma27/pkgs/hydra/overlay.nix
@@ -1,7 +1,16 @@
 self: super:
 
-{
-  hydra = super.hydra.overrideAttrs (_: {
-    patches = [ ./hydra-restricted-eval.patch ];
-  });
-}
+let
+
+  hydraBase = super.hydra.override {
+    nix = super.nixUnstable;
+  };
+
+in
+
+  {
+    hydra = hydraBase.overrideAttrs (_: {
+      patches = [ ./hydra-restricted-eval.patch ];
+      doCheck = false;
+    });
+  }